Make mypy happy
lol
