+++ allennlp/training/__init__.py
@@ -1,7 +1,5 @@
 from allennlp.training.checkpointer import Checkpointer
 from allennlp.training.no_op_trainer import NoOpTrainer
 from allennlp.training.callbacks import TrainerCallback
-from allennlp.training.trainer import (
-    Trainer,
-    GradientDescentTrainer,
-)
+from allennlp.training.trainer import Trainer
+from allennlp.training.gradient_descent_trainer import GradientDescentTrainer
